reworked error page
