+++ frontend/src/containers/FilmCardContainer.jsx
@@ -8,7 +8,7 @@ import FilmCard from '../pages/FilmCard';
 import { addToListAndSave, changeFilm, favoriteIconPush } from '../store/actions/complexFilmActions';
 import { addToTemporary, removeNotification } from '../store/actions/filmActions';
 
-const FilmCardContainer = ({ film, addToListAndSave, addToTemporary, changeFilm, error, notification, removeNotification, favorites, favoriteIconPush }) => {
+const FilmCardContainer = ({ film, addToListAndSave, addToTemporary, changeFilm, notification, removeNotification, favorites, favoriteIconPush }) => {
     const handleChangeFilm = useCallback(() => {
         changeFilm();
     }, [changeFilm]);
@@ -35,8 +35,7 @@ const FilmCardContainer = ({ film, addToListAndSave, addToTemporary, changeFilm,
     return (
         !film 
             ? <Spinner /> 
-            : <FilmCard error={error} 
-                        notify={notification} 
+            : <FilmCard notify={notification}
                         removeNotification={removeNotification} 
                         film={film} 
                         addToTemporary={handleMoveFilmToTemporary} 
@@ -49,7 +48,6 @@ const FilmCardContainer = ({ film, addToListAndSave, addToTemporary, changeFilm,
 };
 
 FilmCardContainer.propTypes = {
-    error: PropTypes.string,
     notification: PropTypes.object,
     film: PropTypes.object,
     favoritesList: PropTypes.object,
@@ -62,7 +60,6 @@ FilmCardContainer.propTypes = {
 
 const mapStateToProps = ({ filmReducer }) => ({
     favorites: filmReducer.favorites,
-    error: filmReducer.error,
     film: filmReducer.film,
     notification: filmReducer.notification
 });
